Finished api calls
Added:
 - Author:
  - getbyid Get
  - register Post
  - login Post
  - update Put
  - delete Delete
 - Category
  - Create Post
  - getByid Get
  - Update Put
  - delete Delete
 - Post
  - Create Post
  - getByid Get
  - Update Put
  - delete Delete

+++ backend/src/models/RouterAPI.ts
@@ -1,17 +1,24 @@
 import { Router, Request, Response } from 'express';
 import { ICrud } from '../models/interfaces/ICrud';
+import { AuthorResponseDTO } from './dtos';
+import { decodeJwt, JwtPayload } from '../managers/AuthManager';
 
-class RouterAPI<S extends ICrud<M,DTO>, M, DTO> {
+class RouterAPI<S extends ICrud<M, DTO>, M, DTO> {
     protected router = Router();
     protected service: S;
 
     public constructor(service: S) {
         this.service = service;
-
-        this.router.get('/:id', this.getById);
-        this.router.post('/', this.create);
-        this.router.put('/:id', this.updateById);
-        this.router.delete('/:id', this.deleteById);
+    }
+    protected checkAuthorToken = async (req: Request): Promise<AuthorResponseDTO | null> => {
+        const authHeader = req.headers.authorization;
+        if (!authHeader?.startsWith('Bearer ')) {
+            return null;
+        }
+        const token = authHeader.substring(7);
+        const decoded: JwtPayload = await decodeJwt(token);
+        const author: AuthorResponseDTO = decoded.user;
+        return author;
     }
 
     public getById = async (req: Request, res: Response): Promise<Response> => {
@@ -30,37 +37,61 @@ class RouterAPI<S extends ICrud<M,DTO>, M, DTO> {
     }
 
     public create = async (req: Request, res: Response): Promise<Response> => {
+        const requester: AuthorResponseDTO | null = await this.checkAuthorToken(req);
+        if (!requester) {
+            return res.status(401).json({ message: "Authorization token required" });
+        }
+
         const reqPost: M = req.body;
         try {
-            const resPost = await this.service.create(reqPost);
-            if(!resPost)
-                return res.status(500).json({message:"Error on insert"});
+            const resPost = await this.service.create(reqPost, requester);
+            if (!resPost)
+                return res.status(500).json({ message: "Error on insert" });
             return res.status(201).json(resPost);
         } catch (e) {
-            return res.status(500).json({message:e});
+            return res.status(500).json({ message: e });
         }
     }
 
     public updateById = async (req: Request, res: Response): Promise<Response> => {
+        const requester: AuthorResponseDTO | null = await this.checkAuthorToken(req);
+        if (!requester) {
+            return res.status(401).json({ message: "Authorization token required" });
+        }
+
         const id = Number(req.params.id);
         const reqPost: M = req.body;
         try {
-            const resPost = await this.service.updateById(id,reqPost);
-            if(!resPost)
-                return res.status(500).json({message:"Id not found"});
+            const resPost = await this.service.updateById(id, reqPost, requester);
+            if (!resPost)
+                return res.status(500).json({ message: "Id not found" });
             return res.status(200).json(resPost);
         } catch (e) {
-            return res.status(500).json({message:e});
+            return res.status(500).json({ message: e });
         }
     }
 
     public deleteById = async (req: Request, res: Response): Promise<Response> => {
         const id = Number(req.params.id);
+        const requester: AuthorResponseDTO | null = await this.checkAuthorToken(req);
+        if (!requester) {
+            return res.status(401).json({ message: "Authorization token required" });
+        }
+
         try {
-            await this.service.deleteById(id);
-            return res.status(200).json({message:`Id ${id} deleted`});
+            if (id !== requester.id && requester.role !== "ADMINISTRATOR") {
+                return res.status(401).json({ message: "Unauthorized action" });
+            }
+            const deleted = await this.service.deleteById(id);
+
+            if (!deleted) {
+                return res.status(404).json({ message: "Id not found" });
+            }
+
+            return res.status(200).json({ message: `Id ${id} deleted` });
+
         } catch (e) {
-            return res.status(500).json({message:e});
+            return res.status(500).json({ message: e });
         }
     }
